+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 - `cover_art_url`: Add `cover_art_url` field to albums and singletons which is used by
   `beets>=2.4.0` for artwork resolution.
 
+### Updated
+
+- `genre`: Update musicbrainz genre list
+
 ## [0.23.0] 2025-09-20
 
 ### Added